# docs/swagger.yaml
definitions:
  request.CheckRequest:
    description: The request payload for checking application updates.
    properties:
      installerPackageName:
        description: |-
          InstallerPackageName is the package name of the installer.
          @Required
        example: apk
        type: string
      packageName:
        description: |-
          PackageName is the unique identifier of the application package.
          @Required
        example: com.test.example
        type: string
      versionCode:
        description: |-
          BuildNumber is the current version build of the application.
          Must be greater than 0.
          @Required
        example: 2
        type: integer
      versionName:
        description: |-
          BuildNumber is the current version build of the application.
          Must be greater than 0.
          @Required
        example: 2.0.0
        type: string
    required:
    - installerPackageName
    - packageName
    - versionName
    type: object
  response.CheckResponse:
    description: holds the overall update information for an application.
    properties:
      latestVersion:
        allOf:
        - $ref: '#/definitions/response.VersionInfo'
        description: |-
          LatestVersion contains the details of the latest version if an update is available.
          This field is omitted when UpdateType is 'none'.
      updateType:
        description: |-
          UpdateType indicates the type of update required ('hard', 'soft', or 'none').
          When UpdateType is 'none', the LatestVersion field will not be present in the response.
        type: string
    type: object
  response.StatusResponse:
    properties:
      message:
        type: string
    type: object
  response.VersionInfo:
    description: holds information about the latest version of the app.
    properties:
      url:
        description: Url is the URL where the latest version can be downloaded from.
        type: string
      versionCode:
        description: VersionCode represents the version build of the application.
        type: integer
      versionName:
        description: Name is the version of the latest application version.
        type: string
    type: object
info:
  contact: {}
paths:
  /check:
    post:
      consumes:
      - application/json
      description: Checks if there is an update available for a given application
        package and returns the type of update required along with the latest version
        information, if any.
      parameters:
      - description: Check Request - includes details such as package version and
          current version of the application.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success Response - returns the type of update required ('hard',
            'soft', or 'none') and the latest version information, if an update is
            available.
          schema:
            $ref: '#/definitions/response.CheckResponse'
        "400":
          description: Bad Request - Invalid request body or parameters.
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "404":
          description: Not Found - Update information for the requested package is
            not found.
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "500":
          description: Internal Server Error - Unexpected error occurred during processing.
          schema:
            $ref: '#/definitions/response.StatusResponse'
      summary: Check for application updates
      tags:
      - Application Update
  /download/{filename}:
    get:
      consumes:
      - application/json
      description: Provides the ability to download an APK file stored on the server.
      parameters:
      - description: Filename of the APK to be downloaded
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: APK file
          schema:
            type: file
        "400":
          description: Bad Request - Invalid file type or invalid request format.
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "404":
          description: Not Found - The requested APK file is not found.
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "500":
          description: Internal Server Error - An unexpected error occurred.
          schema:
            $ref: '#/definitions/response.StatusResponse'
      summary: Download APK file
      tags:
      - APK Download
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: This endpoint handles the uploading of an APK file and updates
        the application version information in the database.
      parameters:
      - description: APK File to be uploaded; should follow the specified naming convention.
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successful response indicating the successful processing and
            updating of the application version information.
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "400":
          description: 'Bad Request: Occurs when the request body is invalid, parameters
            are missing or improperly formatted, or if the file does not follow the
            expected naming convention.'
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "404":
          description: 'Not Found: The requested document is not found in the database.'
          schema:
            $ref: '#/definitions/response.StatusResponse'
        "500":
          description: 'Internal Server Error: An error occurred during processing,
            such as a database error, file upload issue, or any other server-side
            problem.'
          schema:
            $ref: '#/definitions/response.StatusResponse'
      summary: Update Application Version
      tags:
      - Application Update
swagger: "2.0"